// apps/admin/src/lib/api/generated/models/pushAnalyticsMetrics.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.51
 */

export interface PushAnalyticsMetrics {
  /** @minimum 0 */
  accepted?: number;
  /** @minimum 0 */
  browser_deliveries?: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  click_rate?: number;
  /** @minimum 0 */
  clicked?: number;
  /** @minimum 0 */
  displayed?: number;
  /** @minimum 0 */
  failed?: number;
  /** @minimum 0 */
  opened?: number;
  /** @minimum 0 */
  published_events?: number;
  /** @minimum 0 */
  queued?: number;
  /** @minimum 0 */
  recipient_accounts?: number;
  /** @minimum 0 */
  retries?: number;
  /** @minimum 0 */
  running?: number;
  /** @minimum 0 */
  skipped?: number;
}